+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - registry: Use GitLab LabKit for correlation and logging
 - registry: Normalize log messages
 - registry/storage/cache/redis: Add Prometheus metrics for Redis cache store
+- registry: Add support for a pprof monitoring server
 
 ## [v2.9.1-gitlab] - 2020-05-05
 ## Added